refactor: move github under client for growth
