+++ packages/github/tests/contract/create_issue.test.mjs
@@ -8,7 +8,7 @@ import {
 	updateIssueHandler,
 } from "testing/src/interceptors/issues_handler.mjs";
 import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
-import { Github} from "../../src/Github.mjs";
+import { Github} from "../../src/client/Github.mjs";
 
 describe("create issues contract", () => {
     process.env.GITHUB_REPOSITORY = "test-org/test-repo"
