+++ packages/github/tests/contract/find_issue.test.mjs
@@ -3,7 +3,7 @@ import {setupServer} from "msw/node";
 import {buildSearchIssues} from "testing/src/builders/issues.mjs";
 import {findIssueFailureHandler, findIssueHandler} from "testing/src/interceptors/issues_handler.mjs";
 import {afterAll, afterEach, beforeAll, describe, expect, it} from "vitest";
-import {Github} from "../../src/Github.mjs";
+import {Github} from "../../src/client/Github.mjs";
 
 describe("search issues contract", () => {
 	process.env.GITHUB_REPOSITORY = "test-org/test-repo";
