+++ packages/github/tests/contract/update_issue.test.mjs
@@ -2,7 +2,7 @@ import {setupServer} from "msw/node";
 import {buildIssues} from "testing/src/builders/issues.mjs";
 import {updateIssueFailureHandler, updateIssueHandler} from "testing/src/interceptors/issues_handler.mjs";
 import {afterAll, afterEach, beforeAll, describe, expect, it} from "vitest";
-import {Github} from "../../src/Github.mjs";
+import {Github} from "../../src/client/Github.mjs";
 
 describe("update issues contract", () => {
 	process.env.GITHUB_REPOSITORY = "test-org/test-repo";
